+++ src/app/app.module.ts
@@ -2,20 +2,20 @@ import {BrowserModule} from '@angular/platform-browser';
 import {NgModule} from '@angular/core';
 
 import {AppComponent} from './app.component';
-import {FirstComponent} from './first/components/first.component';
 import {CartModule} from './cart/cart.module';
+import {FirstModule} from './first/first.module';
 import {ProductsModule} from './product-list/products.module';
 import {SharedModule} from './shared/shared.module';
 
 @NgModule({
   declarations: [
-    AppComponent,
-    FirstComponent
+    AppComponent
   ],
   imports: [
     BrowserModule,
 
     CartModule,
+    FirstModule,
     ProductsModule,
     SharedModule
   ],