+++ src/app/core/models/config-options.model.ts
@@ -1,0 +1,5 @@
+export interface ConfigOptions {
+  id?: string;
+  login?: string;
+  email?: string;
+}