+++ src/app/product-list/products.module.ts
@@ -3,9 +3,13 @@ import {CommonModule} from '@angular/common';
 
 import {ProductListComponent} from './components/product-list.component';
 import {ProductComponent} from '../product/components/product.component';
+import {SharedModule} from '../shared/shared.module';
 
 @NgModule({
-  imports: [CommonModule],
+  imports: [
+    CommonModule,
+    SharedModule
+  ],
   declarations: [
     ProductListComponent,
     ProductComponent