+++ src/app/product/components/product.component.ts
@@ -1,4 +1,4 @@
-import {ChangeDetectionStrategy, Component, EventEmitter, Input, Output} from '@angular/core';
+import {ChangeDetectionStrategy, Component, ElementRef, EventEmitter, Input, Output, ViewChild} from '@angular/core';
 
 import {ProductModel} from '../models/product.model';
 
@@ -10,15 +10,21 @@ import {ProductModel} from '../models/product.model';
 })
 export class ProductComponent {
   @Input() product: ProductModel;
-  @Output() addToCart: EventEmitter<void> = new EventEmitter<void>();
+  @Output() addToCart: EventEmitter<number> = new EventEmitter<number>();
+  @ViewChild('productAmount') productAmountInput: ElementRef<HTMLInputElement>;
 
   onBuy(): void {
-    this.addToCart.emit();
+    this.addToCart.emit(+this.productAmountInput.nativeElement.value);
+    this.productAmountInput.nativeElement.value = '1';
   }
 
   setClasses(): any {
     return {
       'product-container_inactive': !this.product.isAvailable
     };
+  }
+
+  setCrazyTimes(): number {
+    return this.product.isAvailable ? 15 : 0;
   }
 }